PMM-6873 Move functions to template service
