+++ testdata/ia/others/rules.yml
@@ -1,0 +1,28 @@
+---
+rules:
+    - name: other_rules
+      version: 1
+      summary: MySQL connections in use
+      tiers: [anonymous, registered]
+      expr: |-
+        max_over_time(mysql_global_status_threads_connected[5m]) / ignoring (job)
+        mysql_global_variables_max_connections
+        * 100
+        > [[ .threshold ]]
+      params:
+        - name: threshold
+          summary: A percentage from configured maximum
+          unit: '%'
+          type: float
+          range: [0, 100]
+          value: 80
+      for: 5m
+      severity: warning
+      labels:
+        foo: bar
+      annotations:
+        description: |-
+            More than [[ .threshold ]]% of MySQL connections are in use on {{ $labels.instance }}
+            VALUE = {{ $value }}
+            LABELS: {{ $labels }}
+        summary: MySQL too many connections (instance {{ $labels.instance }})
